python_updater/requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
protobuf==4.21.2
requests==2.28.0
scheduler==0.7.4
typeguard==2.13.3
urllib3==1.26.9
